Simplifying the loop treatment
Since we must deal with branches (that aren't loop conditions) exiting
the loops first before removing the other conditional statements and
rearranging the links between basic blocks, it turns out that, after
this first step, every loop will have a unique way in (the loop header)
and a unique way out (the loop latch). This way, the reverse postorder
will always give the same ordering for the blocks inside a loop. From
this observation, plus the fact that we don't modify the cond. branch
for the loop condition (at the loop latch), it follows that we don't
need to temporarily remove the loop backedges.